fix: Address dashboard permission regression in #23586
